+++ default.nix
@@ -1,7 +1,7 @@
 let
   holonixPath = (import ./nix/sources.nix).holonix;
   holonix = import (holonixPath) {
-    holochainVersionId = "v0_1_0-beta-rc_3";
+    holochainVersionId = "v0_1_0-beta-rc_4";
     include = {
       holochainBinaries = true;
       node = false;